renamed seed file, deleted reference seed files and changed package json seed command
